Updated routing, commenting out and excepting unused routes

+++ lib/stormchat_web/controllers/user_controller.ex
@@ -7,12 +7,12 @@ defmodule StormchatWeb.UserController do
 
   action_fallback StormchatWeb.FallbackController
 
-  # TODO: determine if this will be used
-  def index(conn, _params) do
-    users = Users.list_users()
-    render(conn, "index.json", users: users)
-  end
+  # def index(conn, _params) do
+  #   users = Users.list_users()
+  #   render(conn, "index.json", users: users)
+  # end
 
+  # creates a new user and a default current_location for this new user
   def create(conn, %{"user" => user_params}) do
     with {:ok, %User{} = user} <- Users.create_user(user_params) do
       Locations.create_location(%{name: "current_location", user_id: user.id, lat: 0, long: 0})
@@ -24,12 +24,17 @@ defmodule StormchatWeb.UserController do
     end
   end
 
-  # returns the verified user (view doesn't include password_hash)
-  def show(conn, %{"token" => token}) do
+  # returns the appropriate user view based on who's requesting
+  def show(conn, %{"id" => id, "token" => token}) do
     case Phoenix.Token.verify(conn, "auth token", token, max_age: 86400) do
       {:ok, user_id} ->
-        user = Users.get_user!(user_id)
-        render(conn, "show.json", user: user)
+        user = Users.get_user!(id)
+
+        if user_id == id do
+          render(conn, "show.json", user: user)
+        else
+          render(conn, "show_restricted.json", user: user)
+        end
       _else ->
         conn
         |> redirect(to: page_path(conn, :index))
@@ -37,13 +42,13 @@ defmodule StormchatWeb.UserController do
   end
 
   # verifies that the token user matches the user to be updated, then updates
-  def update(conn, %{"user_params" => user_params, "token" => token}) do
+  def update(conn, %{"id" => id, "user_params" => user_params, "token" => token}) do
     case Phoenix.Token.verify(conn, "auth token", token, max_age: 86400) do
       {:ok, user_id} ->
-        user = Users.get_user(user_id)
+        user = Users.get_user(id)
 
-        if user == nil || user.id != user_params["id"] do
-          IO.inspect({:bad_match, user_params["id"], user.id})
+        if user == nil || id != user_id do
+          IO.inspect({:bad_match, id, user_id})
           raise "hax!"
         end
 
@@ -57,10 +62,16 @@ defmodule StormchatWeb.UserController do
   end
 
   # verifies that the token user matches the user to be deleted, then deletes
-  def delete(conn, %{"token" => token}) do
+  def delete(conn, %{"id" => id, "token" => token}) do
     case Phoenix.Token.verify(conn, "auth token", token, max_age: 86400) do
       {:ok, user_id} ->
-        user = Users.get_user!(user_id)
+        user = Users.get_user!(id)
+
+        if user == nil || id != user_id do
+          IO.inspect({:bad_match, id, user_id})
+          raise "hax!"
+        end
+
         with {:ok, %User{}} <- Users.delete_user(user) do
           send_resp(conn, :no_content, "")
         end
